refactor: split CarryOversSection into one-component-per-file
Extract AmountItem, CarryOverRow, and CarryOverList into separate files
so each TSX file contains exactly one React component.
